requirements.txt
flask==3.1.0
python-dotenv==1.0.1
langchain==0.3.18
langchain-openai==0.3.6
python-docx==1.1.2
scikit-learn==1.6.1
flask-cors==5.0.1
